requirements.txt
Flask>=3.0
Flask-Cors>=4.0
Werkzeug>=3.0
requests>=2.31
numpy>=1.26
opencv-python>=4.8
ultralytics>=8.2
pywebview>=5.0
